+++ test/runtests.jl
@@ -7,6 +7,8 @@ using LinearAlgebra
     include("test_golden.jl")
     include("test_convergence.jl")
     include("test_invariants.jl")
+    include("test_symmetry.jl")
+    include("test_contracts.jl")
     include("test_allocations.jl")
     include("test_verification.jl")
     include("test_maxwell_solvers.jl")